src/donations: Expand user's donations query to check for billingEmail (#503)
Currently anonnymous donations are not included in user's profile. Solve this by making necessary queries look for match in either keycloakId, or billingEmail.

Note:
This solution would only work if the logged in user has, either left the email field empty, or has entered the same email as the one belonging to the account.